+++ test/test.js
@@ -9,7 +9,6 @@ const databases = require('./lib/databases').databases;
 const fs = require('fs').promises;
 const logging = require('../lib/logging');
 const ueberdb = require('../index');
-const util = require('util');
 
 const maxKeyLength = 100;
 const randomString = (length = maxKeyLength) => new Randexp(new RegExp(`.{${length}}`)).gen();
@@ -26,18 +25,9 @@ after(async function () {
   }, 5000).unref();
 });
 
-// Returns an object with promisified equivalents of ueberdb.Database methods.
-const promisifyDb = (db) => {
-  const ret = {};
-  const fns = ['init', 'close', 'get', 'getSub', 'findKeys', 'set', 'setSub', 'remove', 'flush'];
-  for (const fn of fns) ret[fn] = util.promisify(db[fn].bind(db));
-  return ret;
-};
-
 describe(__filename, function () {
   let speedTable;
   let db;
-  let pdb;
 
   before(async function () {
     speedTable = new Clitable({
@@ -76,12 +66,11 @@ describe(__filename, function () {
                   ...(readCache ? {} : {cache: 0}),
                   ...(writeBuffer ? {} : {writeInterval: 0}),
                 }, new logging.ConsoleLogger());
-                pdb = promisifyDb(db);
-                await pdb.init();
+                await db.init();
               });
 
               after(async function () {
-                await pdb.close();
+                await db.close();
                 if (dbSettings.filename) await fs.unlink(dbSettings.filename).catch(() => {});
               });
 
@@ -94,22 +83,22 @@ describe(__filename, function () {
                     before(async function () {
                       input = {a: 1, b: new Randexp(/.+/).gen()};
                       key = randomString(maxKeyLength - 1) + (space ? ' ' : '');
-                      await pdb.set(key, input);
+                      await db.set(key, input);
                     });
 
                     it('get(key) -> record', async function () {
-                      const output = await pdb.get(key);
+                      const output = await db.get(key);
                       assert.equal(JSON.stringify(output), JSON.stringify(input));
                     });
 
                     it('get(`${key} `) -> nullish', async function () {
-                      const output = await pdb.get(`${key} `);
+                      const output = await db.get(`${key} `);
                       assert(output == null);
                     });
 
                     if (space) {
                       it('get(key.slice(0, -1)) -> nullish', async function () {
-                        const output = await pdb.get(key.slice(0, -1));
+                        const output = await db.get(key.slice(0, -1));
                         assert(output == null);
                       });
                     }
@@ -119,22 +108,22 @@ describe(__filename, function () {
 
               it('get of unknown key -> nullish', async function () {
                 const key = randomString();
-                assert(await pdb.get(key) == null);
+                assert(await db.get(key) == null);
               });
 
               it('set+get works', async function () {
                 const input = {a: 1, b: new Randexp(/.+/).gen()};
                 const key = randomString();
-                await pdb.set(key, input);
-                const output = await pdb.get(key);
+                await db.set(key, input);
+                const output = await db.get(key);
                 assert.equal(JSON.stringify(output), JSON.stringify(input));
               });
 
               it('set+get with random key/value works', async function () {
                 const input = {testLongString: new Randexp(/[a-f0-9]{50000}/).gen()};
                 const key = randomString();
-                await pdb.set(key, input);
-                const output = await pdb.get(key);
+                await db.set(key, input);
+                const output = await db.get(key);
                 assert.equal(JSON.stringify(output), JSON.stringify(input));
               });
 
@@ -143,11 +132,11 @@ describe(__filename, function () {
                 // TODO setting a key with non ascii chars
                 const key = new Randexp(/([a-z]\w{0,20})foo\1/).gen();
                 await Promise.all([
-                  pdb.set(key, true),
-                  pdb.set(`${key}a`, true),
-                  pdb.set(`nonmatching_${key}`, false),
+                  db.set(key, true),
+                  db.set(`${key}a`, true),
+                  db.set(`nonmatching_${key}`, false),
                 ]);
-                const keys = await pdb.findKeys(`${key}*`, null);
+                const keys = await db.findKeys(`${key}*`, null);
                 assert.deepEqual(keys.sort(), [key, `${key}a`]);
               });
 
@@ -156,80 +145,80 @@ describe(__filename, function () {
                 if (database === 'redis') this.skip(); // TODO: Fix redis.
                 const key = new Randexp(/([a-z]\w{0,20})foo\1/).gen();
                 await Promise.all([
-                  pdb.set(key, true),
-                  pdb.set(`${key}a`, true),
-                  pdb.set(`${key}b`, false),
-                  pdb.set(`${key}b2`, false),
-                  pdb.set(`nonmatching_${key}`, false),
+                  db.set(key, true),
+                  db.set(`${key}a`, true),
+                  db.set(`${key}b`, false),
+                  db.set(`${key}b2`, false),
+                  db.set(`nonmatching_${key}`, false),
                 ]);
-                const keys = await pdb.findKeys(`${key}*`, `${key}b*`);
+                const keys = await db.findKeys(`${key}*`, `${key}b*`);
                 assert.deepEqual(keys.sort(), [key, `${key}a`].sort());
               });
 
               it('remove works', async function () {
                 const input = {a: 1, b: new Randexp(/.+/).gen()};
                 const key = randomString();
-                await pdb.set(key, input);
-                assert.equal(JSON.stringify(await pdb.get(key)), JSON.stringify(input));
-                await pdb.remove(key);
-                assert(await pdb.get(key) == null);
+                await db.set(key, input);
+                assert.equal(JSON.stringify(await db.get(key)), JSON.stringify(input));
+                await db.remove(key);
+                assert(await db.get(key) == null);
               });
 
               it('getSub of existing property works', async function () {
-                await pdb.set('k', {sub1: {sub2: 'v'}});
-                assert.equal(await pdb.getSub('k', ['sub1', 'sub2']), 'v');
-                assert.deepEqual(await pdb.getSub('k', ['sub1']), {sub2: 'v'});
-                assert.deepEqual(await pdb.getSub('k', []), {sub1: {sub2: 'v'}});
+                await db.set('k', {sub1: {sub2: 'v'}});
+                assert.equal(await db.getSub('k', ['sub1', 'sub2']), 'v');
+                assert.deepEqual(await db.getSub('k', ['sub1']), {sub2: 'v'});
+                assert.deepEqual(await db.getSub('k', []), {sub1: {sub2: 'v'}});
               });
 
               it('getSub of missing property returns nullish', async function () {
-                await pdb.set('k', {sub1: {}});
-                assert(await pdb.getSub('k', ['sub1', 'sub2']) == null);
+                await db.set('k', {sub1: {}});
+                assert(await db.getSub('k', ['sub1', 'sub2']) == null);
 
-                await pdb.set('k', {});
-                assert(await pdb.getSub('k', ['sub1', 'sub2']) == null);
-                assert(await pdb.getSub('k', ['sub1']) == null);
+                await db.set('k', {});
+                assert(await db.getSub('k', ['sub1', 'sub2']) == null);
+                assert(await db.getSub('k', ['sub1']) == null);
 
-                await pdb.remove('k');
-                assert(await pdb.getSub('k', ['sub1', 'sub2']) == null);
-                assert(await pdb.getSub('k', ['sub1']) == null);
-                assert(await pdb.getSub('k', []) == null);
+                await db.remove('k');
+                assert(await db.getSub('k', ['sub1', 'sub2']) == null);
+                assert(await db.getSub('k', ['sub1']) == null);
+                assert(await db.getSub('k', []) == null);
               });
 
               it('setSub can modify an existing property', async function () {
-                await pdb.set('k', {sub1: {sub2: 'v'}});
-                await pdb.setSub('k', ['sub1', 'sub2'], 'v2');
-                assert.deepEqual(await pdb.get('k'), {sub1: {sub2: 'v2'}});
+                await db.set('k', {sub1: {sub2: 'v'}});
+                await db.setSub('k', ['sub1', 'sub2'], 'v2');
+                assert.deepEqual(await db.get('k'), {sub1: {sub2: 'v2'}});
 
-                await pdb.setSub('k', ['sub1'], 'v2');
-                assert.deepEqual(await pdb.get('k'), {sub1: 'v2'});
+                await db.setSub('k', ['sub1'], 'v2');
+                assert.deepEqual(await db.get('k'), {sub1: 'v2'});
 
-                await pdb.setSub('k', [], 'v3');
-                assert.equal(await pdb.get('k'), 'v3');
+                await db.setSub('k', [], 'v3');
+                assert.equal(await db.get('k'), 'v3');
               });
 
               it('setSub can add a new property', async function () {
-                await pdb.remove('k');
-                await pdb.setSub('k', [], {});
-                assert.deepEqual(await pdb.get('k'), {});
-                await pdb.setSub('k', ['sub1'], {});
-                assert.deepEqual(await pdb.get('k'), {sub1: {}});
-                await pdb.setSub('k', ['sub1', 'sub2'], 'v');
-                assert.deepEqual(await pdb.get('k'), {sub1: {sub2: 'v'}});
+                await db.remove('k');
+                await db.setSub('k', [], {});
+                assert.deepEqual(await db.get('k'), {});
+                await db.setSub('k', ['sub1'], {});
+                assert.deepEqual(await db.get('k'), {sub1: {}});
+                await db.setSub('k', ['sub1', 'sub2'], 'v');
+                assert.deepEqual(await db.get('k'), {sub1: {sub2: 'v'}});
 
-                await pdb.remove('k');
-                await pdb.setSub('k', ['sub1', 'sub2'], 'v');
-                assert.deepEqual(await pdb.get('k'), {sub1: {sub2: 'v'}});
+                await db.remove('k');
+                await db.setSub('k', ['sub1', 'sub2'], 'v');
+                assert.deepEqual(await db.get('k'), {sub1: {sub2: 'v'}});
               });
 
               it('setSub rejects attempts to set properties on primitives', async function () {
                 for (const v of ['hello world', 42, true]) {
-                  await pdb.set('k', v);
-                  assert.rejects(pdb.setSub('k', ['sub'], 'x'), {
+                  await db.set('k', v);
+                  assert.rejects(db.setSub('k', ['sub'], 'x'), {
                     name: 'TypeError',
                     message: /property "sub" on non-object/,
                   });
-                  assert.deepEqual(await pdb.get('k'), v);
+                  assert.deepEqual(await db.get('k'), v);
                 }
               });
 
@@ -253,21 +242,21 @@ describe(__filename, function () {
 
                 const timers = {start: Date.now()};
 
-                for (let i = 0; i < count; ++i) promises[i] = pdb.set(key + i, input);
-                promises[count] = pdb.flush();
+                for (let i = 0; i < count; ++i) promises[i] = db.set(key + i, input);
+                promises[count] = db.flush();
                 await Promise.all(promises);
                 timers.set = Date.now();
 
-                for (let i = 0; i < count; ++i) promises[i] = pdb.get(key + i);
+                for (let i = 0; i < count; ++i) promises[i] = db.get(key + i);
                 await Promise.all(promises);
                 timers.get = Date.now();
 
-                for (let i = 0; i < count; ++i) promises[i] = pdb.findKeys(key + i, null);
+                for (let i = 0; i < count; ++i) promises[i] = db.findKeys(key + i, null);
                 await Promise.all(promises);
                 timers.findKeys = Date.now();
 
-                for (let i = 0; i < count; ++i) promises[i] = pdb.remove(key + i);
-                promises[count] = pdb.flush();
+                for (let i = 0; i < count; ++i) promises[i] = db.remove(key + i);
+                promises[count] = db.flush();
                 await Promise.all(promises);
                 timers.remove = Date.now();